api/pyproject.toml
[project]
name = "themes-api"
version = "0.1.0"
description = "REST API for stock themes analytics — regime classification, drift detection, and theme ranking"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "httpx>=0.27"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
